refactor(llm-agent): move LLM call strategies from llm-agent-server
Move NonStreamingLlmCallStrategy, StreamingLlmCallStrategy, and FallbackLlmCallStrategy
from llm-agent-server/src/smart-agent/policy/ to llm-agent/src/policy/. Replace source
files with re-export shims to maintain backward compatibility.
